dart ndt7 client implemented + some bugs fixed

+++ speedtest_mobile/ndt7_client/lib/models/client_response.dart
@@ -42,12 +42,12 @@ class Data {
       };
 
   factory Data.fromJson(Map<String, dynamic> json) => Data(
-        json['ElapsedTime'] as double,
+        json['ElapsedTime'] as int,
         json['NumBytes'] as int,
         json['MeanClientMbps'] as double,
       );
 
-  final double elapsedTime;
+  final int elapsedTime;
   final int numBytes;
   final double meanClientMbps;
 }
